refactor(utils): move test files to __tests__/ directories
